Deadlock_Overlay/requirements.txt
#list of dependencies

screeninfo
opencv-python
pillow
numpy
mss

